Move currency toggle to its own component.
